Create README.md
A platform for people to showcase their talent to the world. Choose your passion from 31 different communities and update what's new to get ranked Globally, Nationally and Locally. Use social network in a completely different way.


Features on the Rankkk app include:

· Connect to people of same and different choices across the world and discover what they are updating in the globe section.

· Update videos and images to showcase your talent and skills across the world.

· Get ranked globally, nationally and locally.

· Get a like, dislike and comment option to express your views.

· Get a notification when a user likes or comment on your update.

· Appreciate who is leading the board in the community section.

· Write a description of you and your passion in the bio section.



The Rankkk app helps you more than staying connected to people, it can be a medium to convert dreams into possibilities.



The best version is yet to come. Stay updated.



For feedback on the app mail to contact@rankkk.com or tell us at http://rankkk.com/contact.php.

Rankkk is available for users of age 13 and over.